refactor(xarm): one grasping stack, hardware by default
Collapses xarm-grasp-sim, xarm-grasp-sim-graspgenx and xarm-perception into
xarm-grasp and xarm-grasp-graspgenx. The two differ only in which GraspGenSpec
provider they compose; everything else is shared, including the arm-versus-sim
split, which is decided at import time the way mujoco_if_sim already does it.

Hardware is the default and --simulation swaps the hardware adapter, the base
pose, the camera, the detector backends and the home pose. That also gives the
hardware path a coordinator, which xarm-perception never had -- it could plan
but had nothing to drive the arm with -- plus the voxel map, which is now always
composed rather than sim-only.
