+++ dimos/robot/manipulators/xarm/blueprints/agentic.py
@@ -25,12 +25,8 @@ from dimos.robot.manipulators.common.agent_prompts import (
     MANIPULATION_AGENT_SYSTEM_PROMPT,
 )
 from dimos.robot.manipulators.xarm.blueprints.basic import xarm7_planner_coordinator
-from dimos.robot.manipulators.xarm.blueprints.perception import xarm_perception
-from dimos.robot.manipulators.xarm.blueprints.simulation import (
-    xarm_grasp_sim,
-    xarm_grasp_sim_graspgenx,
-    xarm_perception_sim,
-)
+from dimos.robot.manipulators.xarm.blueprints.grasp import xarm_grasp, xarm_grasp_graspgenx
+from dimos.robot.manipulators.xarm.blueprints.simulation import xarm_perception_sim
 
 xarm7_planner_coordinator_agent = autoconnect(
     xarm7_planner_coordinator,
@@ -39,29 +35,23 @@ xarm7_planner_coordinator_agent = autoconnect(
     McpClient.blueprint(system_prompt=BASE_MANIPULATION_AGENT_SYSTEM_PROMPT),
 )
 
-xarm_perception_agent = autoconnect(
-    xarm_perception,
+# The stack already carries a detector, a segmenter and a voxel mapper; adding
+# the agent on top packs them into one worker, where the detector's lazy
+# transformers import fails outright, so give them room.
+xarm_grasp_agent = autoconnect(
+    xarm_grasp,
     McpServer.blueprint(),
     McpClient.blueprint(system_prompt=MANIPULATION_AGENT_SYSTEM_PROMPT),
-)
+).global_config(n_workers=6)
+
+xarm_grasp_graspgenx_agent = autoconnect(
+    xarm_grasp_graspgenx,
+    McpServer.blueprint(),
+    McpClient.blueprint(system_prompt=MANIPULATION_AGENT_SYSTEM_PROMPT),
+).global_config(n_workers=6)
 
 xarm_perception_sim_agent = autoconnect(
     xarm_perception_sim,
     McpServer.blueprint(),
     McpClient.blueprint(system_prompt=MANIPULATION_AGENT_SYSTEM_PROMPT),
 )
-
-# The scene blueprints already carry a detector, a segmenter and a voxel mapper;
-# adding the agent on top packs every one of them into a single worker, where the
-# detector's lazy transformers import fails outright.
-xarm_grasp_sim_agent = autoconnect(
-    xarm_grasp_sim,
-    McpServer.blueprint(),
-    McpClient.blueprint(system_prompt=MANIPULATION_AGENT_SYSTEM_PROMPT),
-).global_config(n_workers=6)
-
-xarm_grasp_sim_graspgenx_agent = autoconnect(
-    xarm_grasp_sim_graspgenx,
-    McpServer.blueprint(),
-    McpClient.blueprint(system_prompt=MANIPULATION_AGENT_SYSTEM_PROMPT),
-).global_config(n_workers=6)
