+++ dimos/robot/manipulators/xarm/blueprints/test_teleop.py
@@ -22,7 +22,7 @@ from dimos.manipulation.manipulation_module import (
 )
 from dimos.manipulation.visualization.viser.config import ViserVisualizationConfig
 from dimos.robot.manipulators.xarm.blueprints.basic import xarm7_planner_coordinator
-from dimos.robot.manipulators.xarm.blueprints.perception import xarm_perception
+from dimos.robot.manipulators.xarm.blueprints.grasp import xarm_grasp
 from dimos.robot.manipulators.xarm.blueprints.simulation import xarm_perception_sim
 from dimos.robot.manipulators.xarm.blueprints.teleop import (
     coordinator_teleop_xarm6,
@@ -55,7 +55,7 @@ def _coordinator_tasks(blueprint: Blueprint) -> list[TaskConfig]:
     [
         keyboard_teleop_xarm6,
         keyboard_teleop_xarm7,
-        xarm_perception,
+        xarm_grasp,
         xarm_perception_sim,
     ],
 )
