+++ dimos/robot/test_all_blueprints.py
@@ -65,8 +65,10 @@ SELF_HOSTED_BLUEPRINTS = frozenset(
         "teleop-quest-xarm6",
         "teleop-quest-xarm7",
         "teleop-quest-xarm7-video",
-        "xarm-perception",
-        "xarm-perception-agent",
+        "xarm-grasp",
+        "xarm-grasp-agent",
+        "xarm-grasp-graspgenx",
+        "xarm-grasp-graspgenx-agent",
         "xarm-perception-sim",
         "xarm-perception-sim-agent",
         "xarm7-planner-coordinator",
